Improve automatic placement of video stabilization anchors and reference points
Anchor points are not placed on the limb of an overexposed
solar disc. For both the anchors and reference points, a new
placement criterion is used: the histogram of local brightness
gradient directions must be sufficiently filled (this prevents
placing a point in area containing a single edge, which would
cause the point to jump along it during block matching).